scr/requirements-freeze.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-auth-ldap==5.2.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
frozenlist==1.7.0
GDAL==3.8.4
geoip2==5.1.0
idna==3.10
inflection==0.5.1
invoke==2.2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
maxminddb==2.8.2
multidict==6.6.4
numpy==2.3.2
paramiko==4.0.0
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pypng==0.20220715.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-ldap==3.4.4
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
shapely==2.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
